Add gitignore file and new Jupyter notebook for BoulderingSegment
The commit adds a .gitignore file to specify intentionally untracked files to ignore. It also adds a Jupyter notebook file for BoulderingSegment. The BoulderingSegment notebook includes a new image segmentation model trained on the 'oxford_iiit_pet' dataset and preprocessing functions for custom images to fit as input to the model.